+++ pkg/metal/apis/validation/validation.go
@@ -4,7 +4,7 @@ package validation
 import (
 	"fmt"
 
-	api "github.com/metal-stack/machine-controller-manager-provider-metal/pkg/provider/apis"
+	api "github.com/metal-stack/machine-controller-manager-provider-metal/pkg/metal/apis"
 	corev1 "k8s.io/api/core/v1"
 )
 
@@ -12,19 +12,19 @@ import (
 func ValidateMetalProviderSpec(spec *api.MetalProviderSpec, secrets *corev1.Secret) []error {
 	var allErrs []error
 
-	if "" == spec.Image {
+	if spec.Image == "" {
 		allErrs = append(allErrs, fmt.Errorf("image is required field"))
 	}
-	if "" == spec.Network {
+	if spec.Network == "" {
 		allErrs = append(allErrs, fmt.Errorf("network is required field"))
 	}
-	if "" == spec.Partition {
+	if spec.Partition == "" {
 		allErrs = append(allErrs, fmt.Errorf("partition is required field"))
 	}
-	if "" == spec.Project {
+	if spec.Project == "" {
 		allErrs = append(allErrs, fmt.Errorf("project is required field"))
 	}
-	if "" == spec.Size {
+	if spec.Size == "" {
 		allErrs = append(allErrs, fmt.Errorf("size is required field"))
 	}
 
@@ -35,11 +35,11 @@ func ValidateMetalProviderSpec(spec *api.MetalProviderSpec, secrets *corev1.Secr
 
 func validateSecrets(secret *corev1.Secret) []error {
 	var allErrs []error
-	if ("" == string(secret.Data["metalAPIHMac"])) == ("" == string(secret.Data["metalAPIKey"])) {
-		allErrs = append(allErrs, fmt.Errorf("Either metalAPIHMac or metalAPIKey is required field"))
+	if (string(secret.Data["metalAPIHMac"]) == "") == (string(secret.Data["metalAPIKey"]) == "") {
+		allErrs = append(allErrs, fmt.Errorf("either metalAPIHMac or metalAPIKey is required field"))
 	}
-	if "" == string(secret.Data["metalAPIURL"]) {
-		allErrs = append(allErrs, fmt.Errorf("Secret metalAPIURL is required field"))
+	if string(secret.Data["metalAPIURL"]) == "" {
+		allErrs = append(allErrs, fmt.Errorf("secret metalAPIURL is required field"))
 	}
 	return allErrs
 }
