+++ pkg/metal/core.go
